% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideal.points.dw}
\alias{ideal.points.dw}
\title{Regular DW-NOMINATE scores data.}
\format{\describe{
    \item{\code{cong}}{Congressional Session}
    \item{\code{idno}}{ICPSR ID Number}
    \item{\code{state}}{State Code}
    \item{\code{cd}}{Congressional District Number}
    \item{\code{statenm}}{State Name}
    \item{\code{party}}{Party Code}
    \item{\code{name}}{Name}
    \item{\code{dwnom1}}{1st Dimension Coordinate}
    \item{\code{dwnom2}}{2nd Dimension Coordinate}
    \item{\code{bootse1}}{1st Dimension Bootstrapped Standard Error}
    \item{\code{bootse2}}{2nd Dimension Bootstrapped Standard Error}
    \item{\code{corr12}}{Correlation Between 1st and 2nd Dimension Bootstrapped Estimates}
    \item{\code{logl}}{Log-Likelihood}
    \item{\code{nchoices}}{Number of Votes}
    \item{\code{errors}}{Number of Classification Errors}
    \item{\code{gmp}}{Geometric Mean Probability}
}}
\source{
\url{http://voteview.org/dwnomin_joint_house_and_senate.htm}

\url{ftp://k7moa.com/junkord/HL01113D21_BSSE_12.DTA}

\url{ftp://k7moa.com/junkord/SL01113D21_BSSE_12.DTA}
}
\usage{
ideal.points.dw
}
\description{
Regular DW-NOMINATE scores data.
}
\keyword{datasets}
