% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideal.points.wk31}
\alias{ideal.points.wk31}
\title{Common space DW-NOMINATE scores data (weekly).}
\format{\describe{
    \item{\code{cong}}{Congressional Session}
    \item{\code{idno}}{ICPSR ID Number}
    \item{\code{state}}{State Code}
    \item{\code{cd}}{Congressional District Number}
    \item{\code{statenm}}{State Name}
    \item{\code{party}}{Party Code}
    \item{\code{occupancy}}{ICPSR Occupancy Code}
    \item{\code{lastmeans}}{Last Means of Attaining Office}
    \item{\code{name}}{Name}
    \item{\code{dwnom1}}{1st Dimension Coordinate}
    \item{\code{dwnom2}}{2nd Dimension Coordinate}
    \item{\code{logl}}{Log-Likelihood}
    \item{\code{nchoices}}{Number of Votes}
    \item{\code{errors}}{Number of Classification Errors}
    \item{\code{gmp}}{Geometric Mean Probability}
}}
\source{
\url{http://voteview.org/Weekly_Constant_Space_DW-NOMINATE_Scores.htm}

\url{ftp://k7moa.com/wf1/Weekly_DW-NOMINATE_31_12.DTA}
}
\usage{
ideal.points.wk31
}
\description{
Common space DW-NOMINATE scores data (weekly).
}
\keyword{datasets}
