% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealPoints.R
\name{idealPoints}
\alias{idealPoints}
\title{Ideal Points.}
\usage{
idealPoints(congress = 100, chamber = "house", state = NULL,
  data.version = "common", mcdonald = TRUE, stump = TRUE)
}
\arguments{
\item{congress}{Numeric. Session number.}

\item{chamber}{Character. "house", "senate" or "both" (includes President).}

\item{state}{Character. Two character state postal abbreviation (e.g., "CA") or "US" (President). Lower case OK (e.g, "ca").}

\item{data.version}{Character. DW-NOMINATE flavor: "common", "common.weekly" or "dw".}

\item{mcdonald}{Logical. Include Larry McDonald (GA-D). Sessions 94-98.}

\item{stump}{Logical. Include Bob Stump (AZ-D). Sessions 95-97.}
}
\value{
An R data frame.
}
\description{
Ideal Points.
}
