% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideologicalDispersion.R
\name{ideologicalDispersion}
\alias{ideologicalDispersion}
\title{Compute ideological dispersion.}
\usage{
ideologicalDispersion(congress = 100, chamber = "house", state = NULL,
  data.version = "common", dimension = "both", wt2D = TRUE,
  mcdonald = TRUE, stump = TRUE)
}
\arguments{
\item{congress}{Numeric. Session number.}

\item{chamber}{Character. "house", "senate" or "both".}

\item{state}{Character. Two character state postal abbreviation (e.g., "CA") or "US" (President). Lower case OK (e.g, "ca").}

\item{data.version}{Character. DW-NOMINATE flavor: "common", "common.weekly" or "dw".}

\item{dimension}{Character. "first", "second" or "both".}

\item{wt2D}{Logical. Weight second dimension.}

\item{mcdonald}{Logical. Include Larry McDonald (GA-D).}

\item{stump}{Logical. Include Bob Stump (AZ-D).}
}
\value{
An R dataframe
}
\description{
Compute ideological dispersion.
}
