% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideologicalDistance.R
\name{ideologicalDistance}
\alias{ideologicalDistance}
\title{Compute ideological distance.}
\usage{
ideologicalDistance(congress = 100, chamber = "house", state = NULL,
  data.version = "common", dimension = "both", wt2D = TRUE,
  mcdonald = TRUE, stump = TRUE, method2D = "euclidean")
}
\arguments{
\item{congress}{Numeric. Session number.}

\item{chamber}{Character. "house", "senate" or "both".}

\item{state}{Character. Two character state postal abbreviation (e.g., "CA") or "US" (President). Lower case OK (e.g, "ca").}

\item{data.version}{Character. DW-NOMINATE flavor: "common", "common.weekly" or "dw".}

\item{dimension}{Character. "first", "second" or "both".}

\item{wt2D}{Logical. Weight second dimension.}

\item{mcdonald}{Logical. Include Larry McDonald (GA-D).}

\item{stump}{Logical. Include Bob Stump (AZ-D).}

\item{method2D}{Character. Method of computing 2D distance: "euclidean" or "manhattan".}
}
\value{
An R dataframe
}
\description{
Compute ideological distance.
}
