% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideologicalOverlap.R
\name{ideologicalOverlap}
\alias{ideologicalOverlap}
\title{Compute ideological overlap.}
\usage{
ideologicalOverlap(congress = 100, chamber = "house", state = NULL,
  data.version = "common", mcdonald = TRUE, stump = TRUE,
  wt2D = FALSE, overlap.trim = 0)
}
\arguments{
\item{congress}{Numeric. Session number.}

\item{chamber}{Character.}

\item{state}{Character. Two character state postal abbreviation (e.g., "CA") or "US" (President). Lower case OK (e.g, "ca").}

\item{data.version}{Character. DW-NOMINATE flavor: "common", "common.weekly" or "dw".}

\item{mcdonald}{Logical. Include Larry McDonald (GA-D).}

\item{stump}{Logical. Include Bob Stump (AZ-D).}

\item{wt2D}{Logical. Weight second dimension. Default is FALSE.}

\item{overlap.trim}{Numeric. Sets the smaller quantile to trim the range of scores. Must be on [0, 0.5). "0" uses minimum Republican score and maximum Democrat score.}
}
\value{
An R dataframe
}
\description{
Compute ideological overlap.
}
