% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapEndPoints.R
\name{overlapEndPoints}
\alias{overlapEndPoints}
\title{Legislators at endpoints of overlap interval (1st Dimension).}
\usage{
overlapEndPoints(congress = 100, chamber = "house",
  data.version = "common", mcdonald = TRUE, stump = TRUE)
}
\arguments{
\item{congress}{Numeric. Session number.}

\item{chamber}{Character. "house", "senate" or "both".}

\item{data.version}{Character. DW-NOMINATE flavor: "common", "common.weekly" or "dw".}

\item{mcdonald}{Logical. Include Larry McDonald (GA-D).}

\item{stump}{Logical. Include Bob Stump (AZ-D).}
}
\value{
An R data frame. NA indicates no overlap.
}
\description{
Most liberal Republican (lowest score) and most conservative Democrat (highest score).
}
