% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarization1D.R
\name{polarization1D}
\alias{polarization1D}
\title{1D Plot.}
\usage{
polarization1D(congress = 100, chamber = "house", state = NULL,
  data.version = "common", mcdonald = TRUE, stump = TRUE,
  alpha = 1/3, ...)
}
\arguments{
\item{congress}{Numeric. Session number.}

\item{chamber}{Character. "house", "senate" or "both". "both" includes President.}

\item{state}{Character. Two character state postal abbreviation (e.g., "CA") or "US" (President). Lower case is OK (e.g, "ca").}

\item{data.version}{Character. DW-NOMINATE flavor: "common", "common.weekly" or "dw".}

\item{mcdonald}{Logical. Include Larry McDonald (GA-D).}

\item{stump}{Logical. Include Bob Stump (AZ-D).}

\item{alpha}{Numeric. Alpha level transparency.}

\item{...}{Additional plotting parameters.}
}
\value{
An R plot.
}
\description{
Distribution of scores along 1st dimension.
}
