% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarization2D.v12.R
\name{polarization2D}
\alias{polarization2D}
\title{2D Plot (prototype).}
\usage{
polarization2D(congress = 100, chamber = "house", state = NULL,
  data.version = "common", mcdonald = TRUE, stump = TRUE, alpha = 2/3,
  overlap.trim = 0, wt2D = TRUE, all = TRUE, ...)
}
\arguments{
\item{congress}{Numeric. Session number.}

\item{chamber}{Character. "house", "senate" or "both".}

\item{state}{Character. Two character state postal abbreviation (e.g., "CA") or "US" (President). Lower case OK (e.g, "ca").}

\item{data.version}{Character. DW-NOMINATE flavor: "common", "common.weekly" or "dw".}

\item{mcdonald}{Logical. Include Larry McDonald (GA-D). idno = 14252.}

\item{stump}{Logical. Include Bob Stump (AZ-D). idno = 94454.}

\item{alpha}{Numeric. Alpha level transparency.}

\item{overlap.trim}{Numeric. Sets the smaller quantile to trim the range of scores. Must be on [0, 0.5). "0" uses minimum Republican score and maximum Democrat score.}

\item{wt2D}{Logical. Weight second dimension.}

\item{all}{Logical. Show legislator points.}

\item{...}{Additional plotting parameters.}
}
\value{
An R plot.
}
\description{
2D Plot (prototype).
}
\examples{
# polarization2D(chamber = "both")
# polarization2D(state = "CA")
# polarization2D(state = "ny")
}
