% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateCodes.R
\name{stateCodes}
\alias{stateCodes}
\title{State codes and names}
\usage{
stateCodes()
}
\description{
Allows use of the standard two letter postal state abbreviations (e.g., "CA") in other functions.
}
\section{Notes}{
 The President's state is "US".

 "south" is the eleven states of the Confederacy plus Kentucky and Oklahoma (Nolan McCarty, Keith T. Poole and Howard Rosenthal. 2006. "Polarized America: The Dance of Ideology and Unequal Riches". MIT Press. p. 74).
}

